ocr test complete
